+++ lib/src/delegate/future.dart
@@ -7,20 +7,20 @@ import 'dart:async';
 /// A wrapper that forwards calls to a [Future].
 class DelegatingFuture<T> implements Future<T> {
   /// The wrapped [Future].
-  final Future _future;
+  final Future<T> _future;
 
   DelegatingFuture(this._future);
 
   Stream<T> asStream() => _future.asStream();
 
-  Future catchError(Function onError, {bool test(error)}) =>
+  Future catchError(Function onError, {bool test(Object error)}) =>
     _future.catchError(onError, test: test);
 
-  Future then(onValue(T value), {Function onError}) =>
+  Future/*<S>*/ then/*<S>*/(/*=S*/ onValue(T value), {Function onError}) =>
     _future.then(onValue, onError: onError);
 
   Future<T> whenComplete(action()) => _future.whenComplete(action);
 
-  Future timeout(Duration timeLimit, {void onTimeout()}) =>
+  Future<T> timeout(Duration timeLimit, {void onTimeout()}) =>
     _future.timeout(timeLimit, onTimeout: onTimeout);
 }