+++ lib/src/result/capture_transformer.dart
@@ -7,7 +7,10 @@ import 'dart:async';
 import '../result.dart';
 import 'capture_sink.dart';
 
-/// Use [Result.captureTransformer] instead.
+/// A stream transformer that captures a stream of events into [Result]s.
+///
+/// The result of the transformation is a stream of [Result] values and no
+/// error events. This is the transformer used by [captureStream].
 @Deprecated("Will be removed in async 2.0.0.")
 class CaptureStreamTransformer<T> implements StreamTransformer<T, Result<T>> {
   const CaptureStreamTransformer();